create interface module and move related methods
